Fix weird TS compilation issue
