+++ packages/docusaurus/src/templates/templates.ts
@@ -7,7 +7,8 @@
 
 import * as eta from 'eta';
 import {getBundles} from 'react-loadable-ssr-addon-v5-slorber';
-import type {SSGParams, AppRenderResult} from '../types';
+import type {SSGParams} from '../ssg';
+import type {AppRenderResult} from '../common';
 import type {Manifest} from 'react-loadable-ssr-addon-v5-slorber';
 
 // TODO this is historical server template data
